Update the documentation
I've updated the README to include installation instructions with Hugo
modules which is pretty handy. Also, the recipe section in the README is
removed which I think is getting too big. Maybe, it should be on its own
folder or branch.

All of the listed changes should be indicated in the changelog.