Improve the pagination
As much as I avoid creating custom partials, it seem it is not possible
with this component since it has gotten a bit bigger than I expected.

The pagination component is a modified version of the internal
pagination template but smarter. One of the major change is it will only
apply ellipses if it has more than one page to skip. Also, it is easier
to configure now that it is refactored to be one. Last but not least, it
has comments. :)